pyproject.toml
[project]
name = "pdf-fill"
version = "0.1.0"
description = "MCP server for analyzing and filling documents with text, shapes, and visual elements"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "Pillow>=10.0.0",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.0.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
analysis = ["surya-ocr>=0.8.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[project.scripts]
pdf-fill = "pdf_fill.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_fill"]
